requirements.txt
streamlit
sodapy
pandas
plotly
SQLAlchemy
PyMySQL
python-dotenv